+++ packages/contracts/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./access.js";
 export * from "./address.js";
 export * from "./capability.js";
+export * from "./card.js";
 export * from "./common.js";
 export * from "./context-capsule.js";
 export * from "./execution.js";